+++ sylvia-iot-broker/src/routes/v1/application/mod.rs
@@ -61,25 +61,38 @@ pub fn new_service(scope_path: &str, state: &State) -> Router {
         Router::new()
             .route(
                 "/",
-                routing::post(api::post_application)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_root)),
+                routing::post(api::post_application).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_root,
+                )),
             )
             .route(
                 "/count",
-                routing::get(api::get_application_count)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_count)),
+                routing::get(api::get_application_count).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_count,
+                )),
             )
             .route(
                 "/list",
-                routing::get(api::get_application_list)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_list)),
+                routing::get(api::get_application_list).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_list,
+                )),
             )
             .route(
                 "/{application_id}",
                 routing::get(api::get_application)
                     .patch(api::patch_application)
                     .delete(api::delete_application)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_param)),
+                    .layer(AuthService::new(
+                        state.client.clone(),
+                        auth_uri.clone(),
+                        role_scopes_param,
+                    )),
             )
             .with_state(state.clone()),
     )