+++ sylvia-iot-broker/src/routes/v1/network_route/mod.rs
@@ -61,23 +61,35 @@ pub fn new_service(scope_path: &str, state: &State) -> Router {
         Router::new()
             .route(
                 "/",
-                routing::post(api::post_network_route)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_root)),
+                routing::post(api::post_network_route).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_root,
+                )),
             )
             .route(
                 "/count",
-                routing::get(api::get_network_route_count)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_count)),
+                routing::get(api::get_network_route_count).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_count,
+                )),
             )
             .route(
                 "/list",
-                routing::get(api::get_network_route_list)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_list)),
+                routing::get(api::get_network_route_list).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_list,
+                )),
             )
             .route(
                 "/{route_id}",
-                routing::delete(api::delete_network_route)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_param)),
+                routing::delete(api::delete_network_route).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_param,
+                )),
             )
             .with_state(state.clone()),
     )