+++ sylvia-iot-broker/src/routes/v1/unit/mod.rs
@@ -75,30 +75,46 @@ pub fn new_service(scope_path: &str, state: &State) -> Router {
         Router::new()
             .route(
                 "/",
-                routing::post(api::post_unit)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_root)),
+                routing::post(api::post_unit).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_root,
+                )),
             )
             .route(
                 "/count",
-                routing::get(api::get_unit_count)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_count)),
+                routing::get(api::get_unit_count).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_count,
+                )),
             )
             .route(
                 "/list",
-                routing::get(api::get_unit_list)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_list)),
+                routing::get(api::get_unit_list).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_list,
+                )),
             )
             .route(
                 "/{unit_id}",
                 routing::get(api::get_unit)
                     .patch(api::patch_unit)
                     .delete(api::delete_unit)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_param)),
+                    .layer(AuthService::new(
+                        state.client.clone(),
+                        auth_uri.clone(),
+                        role_scopes_param,
+                    )),
             )
             .route(
                 "/user/{user_id}",
-                routing::delete(api::delete_unit_user)
-                    .layer(AuthService::new(auth_uri.clone(), role_scopes_user)),
+                routing::delete(api::delete_unit_user).layer(AuthService::new(
+                    state.client.clone(),
+                    auth_uri.clone(),
+                    role_scopes_user,
+                )),
             )
             .with_state(state.clone()),
     )