+++ sylvia-iot-data/tests/routes/libs.rs
@@ -104,10 +104,12 @@ pub fn new_state(db_engine: Option<&'static str>) -> TestState {
         Err(e) => panic!("create runtime error: {}", e),
         Ok(runtime) => runtime,
     };
+    let client = reqwest::Client::new();
 
     if db_engine.is_none() {
         return TestState {
             runtime: Some(runtime),
+            client: Some(client.clone()),
             ..Default::default()
         };
     }
@@ -253,6 +255,7 @@ pub fn new_state(db_engine: Option<&'static str>) -> TestState {
         mongodb,
         sqlite,
         routes_state: Some(state),
+        client: Some(client),
         ..Default::default()
     }
 }