+++ sylvia-iot-sdk/tests/middlewares/auth.rs
@@ -47,6 +47,7 @@ fn test_200(context: &mut SpecContext<TestState>) -> Result<(), String> {
     let runtime = state.runtime.as_ref().unwrap();
     let auth_db = state.auth_db.as_ref().unwrap();
     let auth_uri = state.auth_uri.as_ref().unwrap();
+    let client = state.client.as_ref().unwrap();
 
     let result: Result<(), Box<dyn StdError>> = runtime.block_on(async move {
         let now = Utc::now();
@@ -64,7 +65,7 @@ fn test_200(context: &mut SpecContext<TestState>) -> Result<(), String> {
 
     let app = Router::new()
         .route("/", routing::get(test_200_handler))
-        .layer(AuthService::new(auth_uri.clone()));
+        .layer(AuthService::new(client.clone(), auth_uri.clone()));
     let server = match TestServer::new(app) {
         Err(e) => return Err(format!("new server error: {}", e)),
         Ok(server) => server,
@@ -88,10 +89,11 @@ fn test_400(context: &mut SpecContext<TestState>) -> Result<(), String> {
     let state = state.get(STATE).unwrap();
     let runtime = state.runtime.as_ref().unwrap();
     let auth_uri = state.auth_uri.as_ref().unwrap();
+    let client = state.client.as_ref().unwrap();
 
     let app = Router::new()
         .route("/", routing::get(dummy_handler))
-        .layer(AuthService::new(auth_uri.clone()));
+        .layer(AuthService::new(client.clone(), auth_uri.clone()));
     let server = match TestServer::new(app) {
         Err(e) => return Err(format!("new server error: {}", e)),
         Ok(server) => server,
@@ -107,10 +109,11 @@ fn test_401(context: &mut SpecContext<TestState>) -> Result<(), String> {
     let state = state.get(STATE).unwrap();
     let runtime = state.runtime.as_ref().unwrap();
     let auth_uri = state.auth_uri.as_ref().unwrap();
+    let client = state.client.as_ref().unwrap();
 
     let app = Router::new()
         .route("/", routing::get(dummy_handler))
-        .layer(AuthService::new(auth_uri.clone()));
+        .layer(AuthService::new(client.clone(), auth_uri.clone()));
     let server = match TestServer::new(app) {
         Err(e) => return Err(format!("new server error: {}", e)),
         Ok(server) => server,
@@ -129,10 +132,11 @@ fn test_503(context: &mut SpecContext<TestState>) -> Result<(), String> {
     let state = state.get(STATE).unwrap();
     let runtime = state.runtime.as_ref().unwrap();
     let auth_uri = "http://localhost:65535";
+    let client = state.client.as_ref().unwrap();
 
     let app = Router::new()
         .route("/", routing::get(dummy_handler))
-        .layer(AuthService::new(auth_uri.to_string()));
+        .layer(AuthService::new(client.clone(), auth_uri.to_string()));
     let server = match TestServer::new(app) {
         Err(e) => return Err(format!("new server error: {}", e)),
         Ok(server) => server,
@@ -259,6 +263,7 @@ fn before_all_fn(state: &mut HashMap<&'static str, TestState>) -> () {
             auth_db,
             core_svc: Some(core_svc),
             auth_uri,
+            client: Some(reqwest::Client::new()),
             ..Default::default()
         },
     );